move customdropdown into fields-controls
